requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.1.8
decorator==4.4.2
distro==1.9.0
elevenlabs @ git+https://github.com/elevenlabs/elevenlabs-python.git@37cc4954a32c376874a78a6cd772f6d63b2d523d
feedparser==6.0.11
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
jiter==0.10.0
moviepy @ git+https://github.com/Zulko/moviepy.git@3fd700c2d2235f6e03c84f8ee8d844a21e2ad4a2
numpy==2.3.1
openai==1.93.3
pillow==11.3.0
proglog==0.1.12
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
python-dotenv==1.1.1
requests==2.32.4
sgmllib3k==1.0.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
websockets==15.0.1
yt-dlp==2025.6.30
replicate==0.33.0
